Move public-facing views to a separate file
